move client key handler to root
