+++ model.go
@@ -4,7 +4,6 @@ import (
 	"github.com/charmbracelet/bubbles/viewport"
 	tea "github.com/charmbracelet/bubbletea"
 
-	"github.com/marang/emqutiti/clientkeys"
 	"github.com/marang/emqutiti/confirm"
 	"github.com/marang/emqutiti/connections"
 	"github.com/marang/emqutiti/focus"
@@ -19,7 +18,7 @@ import (
 
 // reference history package to avoid unused import warning
 var _ = history.ID
-var _ clientkeys.Handler = (*model)(nil)
+var _ Handler = (*model)(nil)
 var _ message.Model = (*model)(nil)
 
 // Component defines a screen or feature that can participate in the Tea update
